rename file to match refactored class name
