+++ pyomo/contrib/piecewise/transform/outer_representation_gdp.py
@@ -12,7 +12,7 @@
 import pyomo.common.dependencies.numpy as np
 from pyomo.common.dependencies.scipy import spatial
 from pyomo.contrib.fbbt.fbbt import compute_bounds_on_expr
-from pyomo.contrib.piecewise.transform.piecewise_to_gdp_transformation import (
+from pyomo.contrib.piecewise.transform.piecewise_linear_transformation_base import (
     PiecewiseLinearTransformationBase,
 )
 from pyomo.core import Constraint, NonNegativeIntegers, Suffix, Var
